Library/CloudStorage/OneDrive-NTNU/AzureTest/pyproject.toml
[project]
name = "azuretest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "jwt>=1.4.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
]
